make every vital art part a MODULE
